Multiple changes to enable ressuspension
Added a routine to create the bathymetric matrix, corrected the ressuspention routine in the vertical motion kernel. Velocities near the bottom are now adjusted to a logaritmic profile just as is done in the old mohid lagrangian model. Also added a horiontal interpolation method which selects the first available value different than 0. This is important for processes near the bottom. Also corrected bugs in the reading of keywords because a verification was done that included the conversion to an integer and if the value was zero, the model would ignore the keyword and use the default. This happened in the beaching level and for the remove land tracers for example. Added a routine to compute the distance between vertical points (DWZ in MOHIDWater) which is used for the bottom velocity computation, and enabled the model to read a bathymetry field from the netcdfs. Finally, all implementations have been adjusted so as to be able to still work properly.